Make linking with libncurses.so/dylib also bring in -ltinfo
Other platforms arrange for linking with -lncurses to also bring
in -ltinfo without needing the latter to be listed on the linker
command line. We do so too to ease porting other software to
Conda's ncurses. Fixes #44.

On macOS, build libncurses.dylib with -reexport-ltinfo (see the
Darwin linker man page).

With the GNU linker, replace the installed libncurses.so with
a small linker script.

Bump meta.yaml build/number.